Optimize callbacks to reduce noise
