+++ .github/workflows/test.yaml
@@ -48,13 +48,6 @@ jobs:
           format: markdown
           output: both
 
-      - name: Add Coverage PR Comment
-        uses: marocchino/sticky-pull-request-comment@v2
-        if: github.event_name == 'pull_request'
-        with:
-          hide_and_recreate: true
-          path: code-coverage-results.md
-
       - name: Write Coverage to Job Summary
         run: cat code-coverage-results.md >> $GITHUB_STEP_SUMMARY
 
@@ -64,4 +57,4 @@ jobs:
         with:
           report_paths: junit/test-results.xml
           detailed_summary: true
-          comment: true
+          comment: false
